+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.18] - 2025.03.18
+
+### Changed
+
+- Mirroring commands
+    - Support reflection across arbitrary coordinates, not just the world origin (0,0,0). Prompt the user to enter the desired coordinate.
+- Rotation commands
+    - Support rotation around arbitrary coordinates, not just the world origin (0,0,0). Prompt the user to enter the desired center of rotation.
+- `node-ty-levels` NPM package
+    - Expose translation functions.
+
 ## [0.0.17] - 2025.03.15
 
 ### Added